+++ libs/lib.securities/src/main/scala/org/aiotrade/lib/securities/model/Ticker.scala
@@ -28,16 +28,16 @@ object Tickers extends Table[Ticker] {
 
   val time = "time" BIGINT
 
-  val prevClose = "prevClose" FLOAT(12, 2)
-  val lastPrice = "lastPrice" FLOAT(12, 2)
+  val prevClose = "prevClose" FLOAT(12, 3)
+  val lastPrice = "lastPrice" FLOAT(12, 3)
 
-  val dayOpen   = "dayOprn"   FLOAT(12, 2)
-  val dayHigh   = "dayHigh"   FLOAT(12, 2)
-  val dayLow    = "dayLow"    FLOAT(12, 2)
-  val dayVolume = "dayVolume" FLOAT(12, 2)
-  val dayAmount = "dayAmount" FLOAT(12, 2)
+  val dayOpen   = "dayOprn"   FLOAT(12, 3)
+  val dayHigh   = "dayHigh"   FLOAT(12, 3)
+  val dayLow    = "dayLow"    FLOAT(12, 3)
+  val dayVolume = "dayVolume" FLOAT(12, 3)
+  val dayAmount = "dayAmount" FLOAT(12, 3)
 
-  val dayChange = "dayChange" FLOAT(12, 2)
+  val dayChange = "dayChange" FLOAT(12, 3)
 
   val bidAsks = "bidAsks" SERIALIZED(classOf[Array[Float]], 200)
